Add "heartbeat" to sysmon-mqtt
Instead of reporting "0" or "1" in the connected-topic, report "0" or
the current Unix-timestamp (and update it at each iteration). This way
the value can be used as a heartbeat for systems running sysmon-mqtt.
No more need to ping the system from Home Assistant to check if its
still running...

Exclude these heartbeat sensors from the recorder (there's no value in
keeping track of their history).

Furthermore, it's now possible to specify both a "device_class" and
"icon" (to allow for custom heartbeat and bandwidth rx/tx icons).