Final sysmon-mqtt tweaks (for now...)
While starting, a connected-state of -1 is reported until the end of the
second monitoring iteration (at which point the "hearbeat" starts).
Together with some slight tweaks of the exit-trap this allows startup
failures to be easily detected.

As part of Home Assistant discovery, the "heartbeat" sensor does not
expire (which makes it easier to track a stale heartbeat in HA; all
other sensors still expire, so that fact can also still be used to
rigger automations).
Furthermore, include "model" and "sw_version" in the device registry
output (with the hardware-/board-type and uname-output respectively).

Finally, provide an install-script to easily setup a systemd-service.