Fix sysmon-mqtt ShellCheck issues
It appears the line "echo $(((rx-rx_prev[i])/10#$SYSMON_INTERVAL))"
somehow crashes ShellCheck 😵 This is not apparent when working with it
in VS Code (no error appears; linting simple stops working) and as such
I've been "running blind" with this file for a while...

Correct all issues the sprung up as a result. I've already added
ShellCheck to the "Lint codebase"-workflow and will also add it to my
local linter-script to prevent future issues like this.

Replaced the offending lines with a (more sensible) fully awk-based
approach. This commit also restores the "ZFS ARC available memory"-
workaround (which was always skipped due to a typo).