Add support for reporting round-trip times
Similar to bandwidth reporting, but instead pinging a set of hosts and
reporting back the average round-trip times.

Some additional minor changes for consistency – most importantly not
reporting zero-values for bandwidth during the first iteration but
simply omitting the output altogether (so as to align behaviour between
bandwidth, APT-check and round-trip times).