Improve linter with latest insights
Mainly those not yet applied to my Home Assistant configuration repo.

Includes "tdg" (TODO get) to scan for errant TODO-comments (and some
nice gold-plating to make it look for "XXX"-comments too 😇).

When running the linter workflow on GitHub, it now makes a distinction
between pushes _in_ pull_requests (where it scans just the files touched
in the pull-requests) and "regular" pushes (where all files in the repo
are scanned).

There's some duplication of code/scripts between this repo and my
dotfiles-repo that I might want to find a better solution for in the
future...