Fix the Docker "Build and push" action
Paths were incorrect 🤐 Also, add a ".dockerignore" to ensure we don't
push irrelevant stuff into the builder.

Finally, while I was at it, I quickly checked the ability to do a
RISC-V build. That works, but requires Alpine's edge-image (and bumping
dependency versions), so not adding that now. Did update the Dockerfile
to be able to pull in the RISC-V version of s6-overlay.