Migrate from Home Assistant configuration repository (#1)
Changes required to split off sysmon-mqtt from https://github.com/thijsputman/home-assistant-config. None of this impacts the shell-script itself; it's all boilerplate and tooling. Compared to how it used to function the home-assistant-config-repo, the following changes were made:

Includes "tdg" (TODO get) to scan for errant TODO-comments (and some nice gold-plating to make it look for "XXX"-comments too 😇).

When running the linter workflow on GitHub, it now makes a distinction between pushes _in_ pull_requests (where it scans just the files touched in the pull-requests) and "regular" pushes (where all files in the repo are scanned).

There's some duplication of code/scripts between this repo and my dotfiles-repo that I might want to find a better solution for in the future...

Finally, while I was at it, I quickly checked the ability to do a RISC-V build. That works, but requires Alpine's edge-image (and bumping dependency versions), so not adding that now. Did update the Dockerfile to be able to pull in the RISC-V version of s6-overlay.