Temporary addition to "Build and push"-workflow
Needs to be further extended in the near future (and moved away from the
"workflow_dispatch"-approach used previously). As that's too much for
now, I've updated the workflow to handle being called from within a
pull-request without updating any "production" tags...

Tags are based on the version string included in "sysmon.sh". If it's a
non-production version (e.g. "1.3.0-dev"), the only tags generated are
the SHA and the full version ("1.3.0-dev").

In case the workflow is dispatched from the "develop"-branch, the "edge"
tag is generated; from the "main"-branch, the "latest" tag is generated
(both regardless of the actual version).

The entire thing should be running based on Git refs/tags instead in the
future...