Last-minute changes and notes for release v1.3.0 (#4)
Use gawk instead of tr to lowercase input. Seems tr in BusyBox might (silently) fail on this. As gawk is required anyway, that seems the most portable way to fix the issue...

Furthermore, update install.sh (and its instructions) to point to the new GitHub-repository and some minor further housekeeping.